backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt>=4.0.0,<4.1.0
python-jose[cryptography]
email-validator
